+++ local.js
@@ -1348,7 +1348,10 @@ local.httpHeaders.register('link',
 				if (attr == 'href') {
 					continue;
 				}
-				if (typeof link[attr] == 'boolean') {
+				if (link[attr] === null) {
+					continue;
+				}
+				if (typeof link[attr] == 'boolean' && link[attr]) {
 					linkParts.push(attr);
 				} else {
 					linkParts.push(attr+'="'+link[attr]+'"');
@@ -3487,13 +3490,23 @@ local.setDispatchWrapper(function(request, response, dispatch) {
 });
 
 // INTERNAL
-// Makes sure response header links are absolute
+// Makes sure response header links are absolute and extracts additional attributes
 var isUrlAbsoluteRE = /(:\/\/)|(^[-A-z0-9]*\.[-A-z0-9]*)/; // has :// or starts with ___.___
 function processResponseHeaders(request, response) {
 	if (response.parsedHeaders.link) {
 		response.parsedHeaders.link.forEach(function(link) {
 			if (isUrlAbsoluteRE.test(link.href) === false)
 				link.href = local.joinRelPath(request.urld, link.href);
+			link.host_domain = local.parseUri(link.href).authority;
+			var peerd = local.parsePeerDomain(link.host_domain);
+			if (peerd) {
+				link.host_user   = peerd.user;
+				link.host_relay  = peerd.relay;
+				link.host_app    = peerd.app;
+				link.host_stream = peerd.stream;
+			} else {
+				link.host_user = link.host_relay = link.host_app = link.host_stream = null;
+			}
 		});
 	}
 }