+++ test/web/helpers.js
@@ -230,7 +230,7 @@ success
 {
   body: "",
   headers: {
-    link: "</>; rel=\"self service via\"; id=\"hosts\", <httpl://_worker.js/>; rel=\"current\", <httpl://test.com/>; rel=\"current http://grimwire.com/rel/test grimwire.com/rel/test grimwire.com\""
+    link: "</>; rel=\"self service via\"; id=\"hosts\", <httpl://_worker.js/>; rel=\"current\"; host_domain=\"_worker.js\", <httpl://test.com/>; rel=\"current http://grimwire.com/rel/test grimwire.com/rel/test grimwire.com\"; host_domain=\"test.com\""
   },
   reason: "ok, no content",
   status: 204